app/pyproject.toml
[tool.poetry]
name = "vall-e-x-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
soundfile = "^0.12.1"
numpy = "^1.26.2"
torch = "^2.1.1"
torchvision = "^0.16.1"
torchaudio = "^2.1.1"
tokenizers = "^0.15.0"
encodec = "^0.1.1"
vocos = "^0.1.0"
langid = "^1.1.6"
unidecode = "^1.3.7"
pyopenjtalk = "^0.3.3"
pypinyin = "^0.49.0"
inflect = "^7.0.0"
cn2an = "^0.5.22"
jieba = "^0.42.1"
eng-to-ipa = "^0.0.2"
sudachipy = "^0.6.7"
sudachidict-core = "^20230927"
nltk = "^3.8.1"
openai-whisper = "^20231117"
phonemizer = "^3.2.1"
matplotlib = "^3.8.2"
psutil = "^5.9.6"
transformers = "^4.35.2"
gradio = "^4.8.0"
fastapi = "^0.104.1"
firebase-admin = "^6.3.0"
azure-storage-blob = "^12.19.0"
azure-identity = "^1.15.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
